a-eye_api/requirements.txt
Flask==2.2.1
keras==2.9.0
numpy==1.23.1
Pillow==9.2.0
